[lldb] [mostly NFC] Large WP foundation: WatchpointResources
This patch is rearranging code a bit to add WatchpointResources to
Process.  A WatchpointResource is meant to represent a hardware
watchpoint register in the inferior process.  It has an address, a
size, a type, and a list of Watchpoints that are using this
WatchpointResource.

This current patch doesn't add any of the features of WatchpointResources
that make them interesting -- a user asking to watch a 24 byte
object could watch this with three 8 byte WatchpointResources.  Or
a Watchpoint on 1 byte at 0x1002 and a second watchpoint on 1 byte
at 0x1003, these must both be served by a single WatchpointResource
on that doubleword at 0x1000 on a 64-bit target, if two hardware
watchpoint registers were used to track these separately, one of
them may not be hit.  Or if you have one Watchpoint on a variable
with a condition set, and another Watchpoint on that same variable
with a command defined or different condition, or ignorecount, both
of those Watchpoints need to evaluate their criteria/commands when
their WatchpointResource has been hit.

There's a bit of code movement to rearrange things in the direction
I'll need for implementing this feature, so I want to start with
reviewing & landing this mostly NFC patch and we can focus on the
algorithmic choices about how WatchpointResources are shared and
handled as they're triggeed, separately.

This patch also stops printing "Watchpoint <n> hit: old value: <x>,
new vlaue: <y>" for Read watchpoints.  I could make an argument for
print "Watchpoint <n> hit: current value <x>" but the current output
doesn't make any sense, and the user can print the value if they
are particularly interested.  Read watchpoints are used primarily
to understand what code is reading a variable.

This patch adds more fallbacks for how to print the objects being
watched if we have types, instead of assuming they are all integral
values, so a struct will print its elements.  As large watchpoints
are added, we'll be doing a lot more of those.

To track the WatchpointSP in the WatchpointResources, I changed
the internal API which took a WatchpointSP and devolved it to a
Watchpoint*, which meant touching several different Process files.
I removed the watchpoint code in ProcessKDP which only reported
that watchpoints aren't supported, the base class does that already.

I haven't yet changed how we receive a watchpoint to identify
the WatchpointResource responsible for the trigger, and identify
all Watchpoints that are using this Resource to evaluate their
conditions etc.  This is the same work that a BreakpointSite needs
to do when it has been tiggered, where multiple Breakpoints may be
at the same address.

There is not yet any printing of the Resources that a Watchpoint
is implemented in terms of ("watchpoint list", or
SBWatchpoint::GetDescription).

"watchpoint set var" and "watchpoint set expression" take a size
argument which was previously 1, 2, 4, or 8 (an enum).  I've
changed this to an unsigned int.  Most hardware implementations
can only watch 1, 2, 4, 8 byte ranges, but with Resources we'll
allow a user to ask for different sized watchpoints and set
them in hardware-expressble terms soon.

I've annotated areas where I know there is work still needed with
LWP_TODO that I'll be working on once this is landed.

I've tested this on aarch64 macOS, aarch64 Linux, and Intel macOS.

https://discourse.llvm.org/t/rfc-large-watchpoint-support-in-lldb/72116

+++ lldb/test/API/commands/watchpoints/watchpoint_count/main.c
@@ -2,8 +2,8 @@
 #include <stdio.h>
 
 int main() {
-  uint8_t x1 = 0;
-  uint16_t x2 = 0;
+  long x1 = 0;
+  long x2 = 0;
 
   printf("patatino\n");
